+++ bindings/pydeck-carto/docs/changelog.rst
@@ -1,8 +1,12 @@
 CHANGELOG
 =========
 
-Releases and associated GitHub PRs for pydeck are documented here.
+Releases and associated GitHub PRs for pydeck-carto are documented here.
 
 0.1 Releases
 ------------
 
+* First release with features to connect to CARTO using pydeck
+* Added support for oauth authentication
+* Added validation functions
+* Enriched the CartoAuth object